fix: Move postTransformers to posts folder
